+++ src/main/java/TImeCAlling/spring/service/user/UserCommandService.java
@@ -1,4 +1,11 @@
 package TImeCAlling.spring.service.user;
 
+import TImeCAlling.spring.web.dto.user.UserRequestDTO;
+import TImeCAlling.spring.web.dto.user.UserResponseDTO;
+
 public interface UserCommandService {
+    
+    UserResponseDTO.UserCreateDTO createUser(UserRequestDTO.UserCreateDTO userCreateDTO);
+    
+    Long deleteUser(Long id);
 }
