+++ src/main/java/TImeCAlling/spring/web/controller/user/UserController.java
@@ -1,28 +1,33 @@
 package TImeCAlling.spring.web.controller.user;
 
 import TImeCAlling.spring.apiPayload.ApiResponse;
-import TImeCAlling.spring.repository.user.UserRepository;
-import TImeCAlling.spring.service.user.UserQueryService;
+import TImeCAlling.spring.service.user.UserCommandService;
 import TImeCAlling.spring.web.dto.user.UserRequestDTO;
 import TImeCAlling.spring.web.dto.user.UserResponseDTO;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/users")
 public class UserController {
     
-    private final UserQueryService userQueryService;
+    private final UserCommandService userCommandService;
     
     @PostMapping
     public ApiResponse<UserResponseDTO.UserCreateDTO> createUser(@RequestBody UserRequestDTO.UserCreateDTO userCreateDTO) {
         
-        UserResponseDTO.UserCreateDTO response = userQueryService.createUser(userCreateDTO);
+        UserResponseDTO.UserCreateDTO response = userCommandService.createUser(userCreateDTO);
         return ApiResponse.onSuccess(response);
+    }
+    
+    @DeleteMapping("/{userId}")
+    public ApiResponse<UserResponseDTO.UserDeleteDTO> deleteUser(@RequestParam Long userId) {
+        
+        Long deleteUserId = userCommandService.deleteUser(userId);
+        return ApiResponse.onSuccess(UserResponseDTO.UserDeleteDTO.builder()
+                .id(deleteUserId)
+                .build());
     }
     
 }
