+++ src/main/java/TImeCAlling/spring/web/dto/user/UserResponseDTO.java
@@ -12,4 +12,13 @@ public class UserResponseDTO {
         
         Long id;
     }
+    
+    @Builder
+    @NoArgsConstructor(access = AccessLevel.PROTECTED)
+    @AllArgsConstructor
+    @Getter
+    public static class UserDeleteDTO {
+        
+        Long id;
+    }
 }
